requirements.txt
openenv-core==0.2.3
torch>=2.5,<3
uvicorn[standard]>=0.30,<1
